% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hstore.R
\name{new_hstore}
\alias{new_hstore}
\title{Create a empty hstore}
\usage{
new_hstore(nr)
}
\arguments{
\item{nr}{Number of records.}
}
\value{
A empty hstore (list of lists) of length \code{nr}.
}
\description{
This function creates an empty list of lists, which can be appended to a
data frame as a hstore column.
}
\seealso{
\code{\link{\%->\%}} to read or edit the resulting data structure.
}

